requirements.txt
aiogram>=3.7,<4.0
python-dotenv>=1.0,<2.0
asyncpg>=0.29,<1.0
redis>=5.0,<6.0
APScheduler>=3.10,<4.0
